Artisan AI/apps/api/requirement.txt
annotated-types==0.7.0
anyio==4.10.0
backports.tarfile==1.2.0
build==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
docstring_parser==0.17.0
dulwich==0.24.1
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.31.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.38.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
more-itertools==10.8.0
msgpack==1.1.1
numpy==2.3.3
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pbs-installer==2025.9.2
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry==2.2.0
poetry-core==2.2.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
RapidFuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tomlkit==0.13.3
trove-classifiers==2025.9.11.17
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
vertexai==1.71.1
virtualenv==20.34.0
websockets==15.0.1
zipp==3.23.0
zstandard==0.25.0
